requirements.txt
Flask-SQLAlchemy
flask-login
flask
Click
dnspython
eventlet
Flask-WTF
greenlet
itsdangerous
Jinja2
MarkupSafe
monotonic
six
Werkzeug==2.0.0
WTForms==2.2.1
python-socketio==4.6.0
python-engineio==3.13.2
Flask-SocketIO==2.3.0
pymysql
cryptography
requests
pillow